Split OpTable.hs: move XML parser to Flexdis86.OpTable.Parse
