+++ flexdis86.cabal
@@ -48,6 +48,7 @@ library
     Flexdis86.DefaultParser
     Flexdis86.Disassembler
     Flexdis86.OpTable
+    Flexdis86.OpTable.Parse
     Flexdis86.Trie
   ghc-options: -Wall -fno-ignore-asserts -O2
