+++ src/Flexdis86/OpTable.hs
@@ -1,32 +1,27 @@
 {- |
 Module      :  $Header$
-Copyright   :  (c) Galois, Inc 2013-2016
-Maintainer  :  jhendrix@galois.com
+Copyright   :  (c) Galois, Inc 2013-2026
+Maintainer  :  langston.barrett@gmail.com
 
-This declares the parser for optable.xml file, which is used to define the
-instruction set.
+Types describing x86 instruction definitions, and utilities for working
+with them at runtime.  XML parsing lives in "Flexdis86.OpTable.Parse".
 -}
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DeriveLift #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE PatternGuards #-}
 module Flexdis86.OpTable
-  ( -- * Primitive types.
-    CPURequirement(..)
+  ( -- * Primitive types
+    Mode(..)
+  , CPURequirement(..)
   , Vendor(..)
   , ModeLimit(..)
     -- * Def
-  , Def
+  , Def(..)
   , defMnemonic
   , defMnemonicSynonyms
   , defVendor
   , defCPUReq
   , modeLimit
   , defSupported
-  , Mode(..)
   , defMode
   , reqAddrSize
   , OperandSizeConstraint(..)
@@ -41,165 +36,32 @@ module Flexdis86.OpTable
   , defOperands
   , x64Compatible
   , vexPrefixes
-    -- * Parsing defs
-  , parseOpTable
+  , supportedCPUReqs
+    -- * Operand lookup
   , lookupOperandType
   ) where
 
 import qualified Control.DeepSeq as DS
 import qualified Control.Monad.Fail as MF
-import           Control.Monad (forM_, unless, when)
-import           Control.Monad.State (MonadState(..), execStateT, gets)
-import           Data.Bits ((.&.), (.|.), shiftR, shiftL)
+import           Data.Bits ((.&.))
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Char8 as BSC
-import           Data.Char
-import           Data.Containers.ListUtils (nubOrd)
-import qualified Data.List as List
 import qualified Data.Map.Strict as Map
-import           Data.Maybe
-import           Data.Word
-import           GHC.Generics
+import           Data.Word (Word8)
+import           GHC.Generics (Generic)
 import           Instances.TH.Lift ()
 import           Language.Haskell.TH.Syntax (Lift)
 import           Lens.Micro (Lens', lens, (^.))
-import           Lens.Micro.Mtl ((%=), (.=), (?=), use)
-import           Numeric (readDec, readHex)
-import           Text.XML.Light
 
 import           Flexdis86.Operand
 import           Flexdis86.Register
 import           Flexdis86.Segment
 import           Flexdis86.Sizes
 
-qname :: String -> QName
-qname nm = QName { qName = nm, qURI = Nothing, qPrefix = Nothing }
-
-ppQName :: QName -> String
-ppQName qnm = qName qnm
-
-data ElemState = ElemState { esName :: QName
-                           , esContent :: [Content]
-                           , esLine :: Maybe Line
-                           }
-
-mkElemState :: Element -> ElemState
-mkElemState e = ElemState { esName = elName e
-                          , esContent = elContent e
-                          , esLine = elLine e
-                          }
-
-newtype ElemParser a = EP { unEP :: ElemState -> Either String (a,ElemState) }
-
-instance Functor ElemParser where
-  fmap f m = EP (\s -> case unEP m s of
-                         Left e -> Left e
-                         Right (v,s') -> Right (f v, s'))
-
-instance Applicative ElemParser where
-  pure v = EP $ \s -> Right (v,s)
-  m <*> h = EP $ \s -> do
-    (f,s1) <- unEP m s
-    (x,s2) <- unEP h s1
-    return (f x, s2)
-
-instance Monad ElemParser where
-  return = pure
-  m >>= h = EP $ \s -> do (v,s') <- unEP m s
-                          unEP (h v) s'
-#if !(MIN_VERSION_base(4,13,0))
-  fail = MF.fail
-#endif
-
-instance MF.MonadFail ElemParser where
-  fail e = EP $ \s -> Left $ show (esLine s) ++ ": " ++  e
-
-instance MonadState ElemState ElemParser where
-  get = EP (\s -> Right (s,s))
-  put v = EP (\_ -> Right ((),v))
-
-runWithElement :: ElemParser a -> Element -> ElemParser a
-runWithElement m e = do
-  s <- get
-  put (mkElemState e)
-  v <- m
-  put s
-  return v
-
-runElemParser :: ElemParser a -> Element -> Either String a
-runElemParser m e = fmap fst (unEP m (mkElemState e))
-
-checkTag :: String -> ElemParser ()
-checkTag nm = do
-  enm <- gets esName
-  when (enm /= qname nm) $
-    fail $ "Unexpected tag " ++ ppQName enm ++ " in element."
-
-getNextElt :: ElemParser (Maybe Element)
-getNextElt = do
-  s <- get
-  case esContent s of
-    [] -> return Nothing
-    (h:r) -> do
-      put s { esContent = r }
-      case h of
-        Elem e -> return (Just e)
-        Text d | dropWhile isSpace (cdData d) == "" -> getNextElt
-        t -> fail $ "Unexpected non-whitespace text: " ++ show t
-
-pushElt :: Element -> ElemParser ()
-pushElt e = do
-  s <- get
-  put s { esContent = Elem e : esContent s }
-
-next :: ElemParser a -> ElemParser a
-next p = do
-  me <- getNextElt
-  case me of
-    Just e -> runWithElement p e
-    Nothing -> fail "Failed to find next element."
-
-opt :: String -> ElemParser a -> ElemParser (Maybe a)
-opt nm p = do
-  me <- getNextElt
-  case me of
-    Nothing -> return Nothing
-    Just e
-      | elName e == qname nm ->
-        fmap Just $ runWithElement p e
-      | otherwise -> pushElt e >> return Nothing
-
-checkEnd :: ElemParser ()
-checkEnd = do
-  me <- getNextElt
-  case me of
-    Nothing -> return ()
-    Just e -> pushElt e >> fail "Expected end of elements."
-
-remainingElts :: ElemParser a -> ElemParser [a]
-remainingElts p = go []
-  where go r = do
-          me <- getNextElt
-          case me of
-            Nothing -> return (reverse r)
-            Just e -> do
-              v <- runWithElement p e
-              go (v:r)
-
-asText :: ElemParser String
-asText = do
-  s <- get
-  case esContent s of
-    [Text d] | cdVerbatim d == CDataText -> return (cdData d)
-    x -> fail (unlines [ "asText found non-text.", show x])
-
-required_text :: String -> ElemParser String
-required_text nm = next $ checkTag nm >> asText
-
 ------------------------------------------------------------------------
--- Misc
+-- Mode
 
--- | Mode effect on instruction semantecs.
+-- | Mode effect on instruction semantics.
 data Mode
      -- | Default operand size is 64 bits in x64 mode.
    = Default64
@@ -209,17 +71,10 @@ data Mode
 
 instance DS.NFData Mode
 
--- | Parse a mode for the instruction.
-parse_mode :: ElemParser (Maybe Mode)
-parse_mode = opt "mode" $ do
-  txt <- asText
-  case txt of
-    "def64" -> return Default64
-    "inv64" -> return Invalid64
-    _ -> fail $ "Invalid mode: " ++ show txt
+------------------------------------------------------------------------
+-- CPURequirement
 
--- | Defines which features are required for instruction
--- to be supported.
+-- | Defines which features are required for instruction to be supported.
 data CPURequirement
      -- | This is a base instruction.
    = Base
@@ -242,65 +97,24 @@ data CPURequirement
    | AESNI
      -- | SHA extension
    | SHA
-
    | AVX    -- ^ Advanced vector extensions
-
-   -- | Bit manipulation instructions
-   | BMI2
-
-   -- | Multi-precision add-carry instructions
-   | ADX
+   | BMI2   -- ^ Bit manipulation instructions
+   | ADX    -- ^ Multi-precision add-carry instructions
   deriving (Eq, Generic, Lift, Ord, Show)
 
 instance DS.NFData CPURequirement
 
-insClassMap :: Map.Map String CPURequirement
-insClassMap = Map.fromList
-  [ (,) "X87"   X87
-  , (,) "X87 UNDOC" X87_undocumented
-  , (,) "smx"   SMX
-  , (,) "sse"   SSE
-  , (,) "sse2"  SSE2
-  , (,) "sse3"   SSE3
-  , (,) "sse3 atom" SSE3_atom
-  , (,) "sse4.1" SSE4_1
-  , (,) "sse4.2" SSE4_2
-  , (,) "aesni" AESNI
-  , (,) "sha" SHA
-  , (,) "avx" AVX
-  , (,) "bmi2" BMI2
-  , (,) "adx" ADX
-  ]
+------------------------------------------------------------------------
+-- Vendor
 
--- | Returns a CPU requirement, possible given a requirement from the outer
--- constext.
-parse_CPURequirement :: CPURequirement -> ElemParser CPURequirement
-parse_CPURequirement c = do
-  fmap (fromMaybe c) $ opt "class" $ do
-    txt <- asText
-    case Map.lookup txt insClassMap of
-      Nothing -> fail $ "Unknown instruction class: " ++ show txt
-      Just v -> do
-        when (c /= Base) $ do
-          fail $ "Instruction class already defined " ++ show (c,v)
-        return v
-
--- | Defines whether instuction is vendor specific.
+-- | Defines whether instruction is vendor specific.
 data Vendor = AMD | Intel
   deriving (Eq, Generic, Lift, Ord, Show)
 
 instance DS.NFData Vendor
 
-parse_vendor :: Maybe Vendor -> ElemParser (Maybe Vendor)
-parse_vendor v = do
-  fmap (maybe v Just) $ opt "vendor" $ do
-    when (isJust v) $ do
-      fail $ "Vendor already defined."
-    txt <- asText
-    case txt of
-      "amd" -> return $ AMD
-      "intel" -> return $ Intel
-      _ -> fail $ "Unknown vendor: " ++ show txt
+------------------------------------------------------------------------
+-- ModeLimit
 
 -- | Indicates restrictions on which mode instruction may run in.
 data ModeLimit
@@ -315,143 +129,8 @@ instance DS.NFData ModeLimit
 valid64 :: ModeLimit -> Bool
 valid64 m = m == AnyMode || m == Only64
 
---------------------------------------------------------------------------------
-
--- | VEX prefixes
--- See "Intel® 64 and IA-32 Architectures Software Developer’s Manual"
--- Vol. 2A 3-3
-data VexSpec = VexSpec
-  { vexUseVVVV    :: Maybe VexUseVVVV   -- ^ Purpose of VVVV, if any.
-  , vexSize       :: VexSize            -- ^ Vector size
-  , vexImpSimd    :: Maybe VexImpSimd   -- ^ Implied SIMD prefix
-  , vexImpOpc     :: Maybe VexImpOpc    -- ^ Implided opcode prefix
-  , vexW          :: Maybe VexW         -- ^ W field
-  , vexAllowShort :: Bool               -- ^ Could we use 2-byte encoding
-  } deriving (Eq,Show)
-
-data VexUseVVVV = VEX_NDS | VEX_NDD | VEX_DDS      deriving (Eq,Show)
-data VexSize    = Vex128 | Vex256                  deriving (Eq,Show)
-data VexImpSimd = Imp0x66 | Imp0xF3 | Imp0xF2      deriving (Eq,Show)
-data VexImpOpc  = Imp0x0F | Imp0x0F38 | Imp0x0F3A  deriving (Eq,Show)
-data VexW       = VexW0 | VexW1                    deriving (Eq,Show)
-
--- | Note: this doesn't do any error checking.
-parseVex :: String -> VexSpec
-parseVex inp =
-  VexSpec { vexUseVVVV = vvvv
-          , vexSize    = size
-          , vexImpSimd = impS
-          , vexImpOpc  = impO
-          , vexW       = w
-          , vexAllowShort = short
-          }
-  where
-  ts0 = chunks inp
-
-  (vvvv,ts1) =
-    case ts0 of
-      f : fs | f == "NDS" -> (Just VEX_NDS, fs)
-             | f == "NDD" -> (Just VEX_NDD, fs)
-             | f == "DDS" -> (Just VEX_DDS, fs)
-      _ -> (Nothing, ts0)
-
-  (size,ts2) =
-    case ts1 of
-      f : fs | f == "128" -> (Vex128, fs)
-             | f == "256" -> (Vex256, fs)
-      _ -> error "VEX specification is missing its size field"
-
-  (impS,ts3) =
-    case ts2 of
-      f : fs | f == "66" -> (Just Imp0x66, fs)
-             | f == "F2" -> (Just Imp0xF2, fs)
-             | f == "F3" -> (Just Imp0xF3, fs)
-      _ -> (Nothing, ts2)
-
-  (impO,ts4) =
-     case ts3 of
-      f : fs | f == "0F"   -> (Just Imp0x0F,   fs)
-             | f == "0F3A" -> (Just Imp0x0F3A, fs)
-             | f == "0F38" -> (Just Imp0x0F38, fs)
-      _ -> (Nothing, ts3)
-
-  (w,ts5) =
-    case ts4 of
-      f : fs | f == "W0" -> (Just VexW0, fs)
-             | f == "W1" -> (Just VexW1, fs)
-      _ -> (Nothing, ts4)
-
-  short =
-    case ts5 of
-      f : _ | f == "WIG" -> True
-      _ -> False
-
-  chunks x = if null x then []
-                       else case break (== '.') x of
-                              (as,_:bs) -> as : chunks bs
-                              _         -> [x]
-
--- | Can this prefix use the short (2 byte) form?
-vexMayBeShort :: VexSpec -> Bool
-vexMayBeShort vp =
-  case vexImpOpc vp of
-    Just Imp0x0F38  -> False
-    Just Imp0x0F3A  -> False
-    _ -> case vexW vp of
-           Just VexW1 -> False
-           _          -> vexAllowShort vp
-
--- | Map a VEX prefix specificaiton for an instruction to the sequences
--- of bytes that will match it.
-vexToBytes :: VexSpec -> [[Word8]]
-vexToBytes vp = short ++ long
-  where
-  long  = [ [0xC4, b1, b2 ] | b1 <- longByte1, b2 <- longByte2 ]
-  short = if vexMayBeShort vp then [ [0xC5, b ] | b <- shortByte ] else []
-
-  shortByte = nubOrd
-              [ field 7 r .|. field 3 vvvv .|. field 2 l .|. pp
-              | r <- [ 0, 1 ], vvvv <- vvvvVals, l <- lVals, pp <- ppVals ]
-  longByte1 = nubOrd
-              [ field 5 rxb .|. m
-              | rxb <- [ 0 .. 7 ], m <- mmmmVals ]
-  longByte2 = nubOrd
-              [ field 7 w .|. field 3 vvvv .|. field 2 l .|. pp
-              | w <- wVals, vvvv <- vvvvVals, l <- lVals, pp <- ppVals ]
-
-  field amt x = shiftL x amt
-
-  vvvvVals = case vexUseVVVV vp of
-               Nothing -> [ 15 ]
-               Just _  -> [ 0 .. 15 ]
-
-  wVals    = case vexW vp of
-               Nothing    -> [ 0, 1 ]
-               Just VexW0 -> [ 0 ]
-               Just VexW1 -> [ 1 ]
-
-  lVals    = case vexSize vp of
-               Vex128 -> [0]
-               Vex256 -> [1]
-
-  ppVals   = case vexImpSimd vp of
-               Nothing      -> [ 0 ]
-               Just Imp0x66 -> [ 1 ]
-               Just Imp0xF3 -> [ 2 ]
-               Just Imp0xF2 -> [ 3 ]
-
-  mmmmVals = case vexImpOpc vp of
-               Nothing -> []
-               Just imp ->
-                 case imp of
-                   Imp0x0F   -> [1]
-                   Imp0x0F38 -> [2]
-                   Imp0x0F3A -> [3]
-
-
-
 ------------------------------------------------------------------------
--- Instruction
+-- OperandSizeConstraint / Def
 
 data OperandSizeConstraint
    = OpSize16
@@ -462,37 +141,39 @@ data OperandSizeConstraint
 instance DS.NFData OperandSizeConstraint
 
 -- | The definition of an instruction.
-data Def = Def  { _defMnemonic :: !BS.ByteString
-                  -- ^ Canonical mnemonic
-                , _defMnemonicSynonyms :: [BS.ByteString]
-                  -- ^ Additional mnemonics, not including
-                  -- the canonical mnemonic. Used e.g. by
-                  -- jump instructions.
-                , _defCPUReq :: CPURequirement
-                  -- ^ XXX: we should be able to combine these.
-                  -- For example, instruction `vaesenc` requires both
-                  -- AES and AVX support.
-                , _defVendor :: Maybe Vendor
-                , _modeLimit :: ModeLimit
-                , _defMode   :: Maybe Mode
-                , _reqAddrSize :: Maybe SizeConstraint
-                , _reqOpSize :: Maybe OperandSizeConstraint
-                , _defPrefix :: [String]
-                  -- ^ List of allowed prefixes.
-                , _requiredPrefix :: Maybe Word8
-                , _defOpcodes :: [Word8]
-                  -- ^ List of opcodes, which should be nonempty for
-                  -- a complete 'Def'.
-                , _requiredMod :: Maybe ModConstraint
-                , _requiredReg :: Maybe Fin8
-                , _requiredRM :: Maybe Fin8
-                , _x87ModRM    :: Maybe Fin64
-                , _vexPrefixes  :: ![ [Word8] ]
-                  -- ^ Allowed VEX prefixes for this instruction.
-                , _defOperands  :: ![OperandType]
-                } deriving (Eq, Generic, Lift, Show)
+--
+-- The constructor and record fields are exported so that
+-- "Flexdis86.OpTable.Parse" can construct values at compile time;
+-- callers outside this package should use the lens accessors below.
+data Def = Def
+  { _defMnemonic         :: !BS.ByteString
+    -- ^ Canonical mnemonic
+  , _defMnemonicSynonyms :: [BS.ByteString]
+    -- ^ Additional mnemonics (e.g., for jump instructions).
+  , _defCPUReq           :: CPURequirement
+  , _defVendor           :: Maybe Vendor
+  , _modeLimit           :: ModeLimit
+  , _defMode             :: Maybe Mode
+  , _reqAddrSize         :: Maybe SizeConstraint
+  , _reqOpSize           :: Maybe OperandSizeConstraint
+  , _defPrefix           :: [String]
+    -- ^ List of allowed prefixes.
+  , _requiredPrefix      :: Maybe Word8
+  , _defOpcodes          :: [Word8]
+    -- ^ List of opcodes (nonempty for a complete 'Def').
+  , _requiredMod         :: Maybe ModConstraint
+  , _requiredReg         :: Maybe Fin8
+  , _requiredRM          :: Maybe Fin8
+  , _x87ModRM            :: Maybe Fin64
+  , _vexPrefixes         :: ![[Word8]]
+    -- ^ Allowed VEX prefix byte sequences for this instruction.
+  , _defOperands         :: ![OperandType]
+  } deriving (Eq, Generic, Lift, Show)
 
 instance DS.NFData Def
+
+------------------------------------------------------------------------
+-- Def lenses
 
 -- | Canonical mnemonic for definition.
 defMnemonic :: Lens' Def BS.ByteString
@@ -567,121 +248,30 @@ vexPrefixes = lens _vexPrefixes (\s v -> s { _vexPrefixes = v })
 defOperands :: Lens' Def [OperandType]
 defOperands = lens _defOperands (\s v -> s { _defOperands = v })
 
+------------------------------------------------------------------------
+-- Filtering
+
 -- | CPU requirements accepted by flexdis86.
 supportedCPUReqs :: [CPURequirement]
-supportedCPUReqs = [Base, SSE, SSE2, SSE3, SSE3_atom, SSE4_1, SSE4_2, X87, AESNI, SHA, AVX, BMI2, ADX]
+supportedCPUReqs =
+  [Base, SSE, SSE2, SSE3, SSE3_atom, SSE4_1, SSE4_2,
+   X87, AESNI, SHA, AVX, BMI2, ADX]
 
--- | Return true if this definition is one supported by flexdis86.
-defSupported :: Def -> Bool
-defSupported d = d^.reqAddrSize /= Just Size16
-                 && d^.defCPUReq `elem` supportedCPUReqs
-                 && x64Compatible d
-
-addOpcode :: MonadState Def m => Word8 -> m ()
-addOpcode c = defOpcodes %= (++ [c])
-
-setDefCPUReq :: MonadState Def m => CPURequirement -> m ()
-setDefCPUReq r = do
-  creq <- use defCPUReq
-  when (creq == Base) $ defCPUReq .= r
-
--- | Parse opcode value
-parse_opcode :: (MonadState Def m, MF.MonadFail m) => String -> m ()
-parse_opcode nm = do
-  case readHex nm of
-    [(v,"")] -> addOpcode v
-    _ | Just r <- List.stripPrefix "/a=" nm
-      -> case r of
-           "16" -> reqAddrSize ?= Size16
-           "32" -> reqAddrSize ?= Size32
-           "64" -> reqAddrSize ?= Size64
-           _ -> fail $ "Unexpected address size: " ++ r
-    _ | Just r <- List.stripPrefix "/m=" nm
-      -> case r of
-           "32"  -> modeLimit .= Only32
-           "64"  -> modeLimit .= Only64
-           "!64" -> modeLimit .= Not64
-           _ -> fail $ "Unexpected mode limit: " ++ r
-    _ | Just r <- List.stripPrefix "/mod=" nm
-      -> case r of
-           "11"  -> requiredMod ?= OnlyReg
-           "!11" -> requiredMod ?= OnlyMem
-           _ -> fail $ "Unexpected mod constraint: " ++ r
-    _ | Just r <- List.stripPrefix "/o=" nm
-      , [(b,"")] <- readDec r
-      , b `elem` [16, 32, 64::Int]
-      -> case r of
-           "16" -> reqOpSize ?= OpSize16
-           "32" -> reqOpSize ?= OpSize32
-           "64" -> reqOpSize ?= OpSize64
-           _ -> fail $ "Unexpected operand size: " ++ r
-
-    _ | Just r <- List.stripPrefix "/reg=" nm
-      , [(b,"")] <- readHex r
-      , Just v <- asFin8 b
-      -> requiredReg ?= v
-    _ | Just r <- List.stripPrefix "/rm=" nm
-      , [(b,"")] <- readHex r
-      , Just v <- asFin8 b
-      -> requiredRM ?= v
-    _ | Just r <- List.stripPrefix "/3dnow=" nm
-      , [(b,"")] <- readHex r
-      -> do setDefCPUReq AMD_3DNOW
-            addOpcode b -- We don't use requiredPrefix here because it is a suffix
-    _ | Just r <- List.stripPrefix "/sse=" nm
-      , [(b,"")] <- readHex r
-      -> do setDefCPUReq SSE
-            requiredPrefix ?= b
-    _ | Just r <- List.stripPrefix "/x87=" nm
-      , [(b,"")] <- readHex r
-      , Just modRM <- asFin64 b
-      -> do setDefCPUReq X87
-            x87ModRM ?= modRM
-            -- FIXME: sjw: HACK to avoid making the parser more complex.  Basically, we
-            -- pretend we want both Reg and R/M
-            requiredRM  ?= maskFin8 b -- bottom 3 bits
-            requiredReg ?= maskFin8 (b `shiftR` 3)
-    -- This is a special hack used for the endbr32 and endbr64 instructions.
-    -- The first byte in their opcodes is parsed as a REP prefix, so we make
-    -- sure that it is present by marking it as a required prefix. See
-    -- Note [x86_64 disassembly] in Flexdis86.Disassembler for more details.
-    _ | Just r <- List.stripPrefix "/reqpfx=" nm
-      , [(b,"")] <- readHex r
-      -> requiredPrefix ?= b
-
-    _ | Just r <- List.stripPrefix "/vex=" nm
-      -> do setDefCPUReq AVX
-            vexPrefixes .= vexToBytes (parseVex r)
-
-    _  ->  fail $ "Unexpected opcode: " ++ show nm
-
-------------------------------------------------------------------------
--- Instruction
-
--- | Return instuction if instruction can be used with x64.
+-- | Return true if this instruction is compatible with 64-bit mode.
 x64Compatible :: Def -> Bool
 x64Compatible d =
   case d^.defOpcodes of
     [b] | null (d^.vexPrefixes) && (b .&. 0xF0 == 0x40) -> False
-    _ -> valid64 (d^.modeLimit)
+    _   -> valid64 (d^.modeLimit)
 
--- | Recognizes form for mnemonics
-isMnemonic :: String -> Bool
-isMnemonic (h:r) = ('a' <= h && h <= 'z' || h == '_') && all isLowerOrDigit r
-isMnemonic [] = False
+-- | Return true if this definition is one supported by flexdis86.
+defSupported :: Def -> Bool
+defSupported d = d^.reqAddrSize /= Just Size16
+              && d^.defCPUReq `elem` supportedCPUReqs
+              && x64Compatible d
 
-isLowerOrDigit :: Char -> Bool
-isLowerOrDigit c = 'a' <= c && c <= 'z' || isDigit c || c == '_'
-
-parse_mnemonics :: ElemParser (BS.ByteString, [BS.ByteString])
-parse_mnemonics = do
-  mnems <- words <$> required_text "mnemonic"
-  forM_ mnems $ \mnem -> do
-    unless (isMnemonic mnem) $
-      fail $ "Invalid mnemonic: " ++ show mnem
-  case BSC.pack <$> mnems of
-    [] -> fail "Empty mnemonic element!"
-    (h:t) -> return (h, t)
+------------------------------------------------------------------------
+-- Operand lookup
 
 operandHandlerMap :: Map.Map String OperandType
 operandHandlerMap = Map.fromList
@@ -763,8 +353,7 @@ operandHandlerMap = Map.fromList
     -- MMX register stored in ModRM.reg
   , (,) "P"    $ RG_MMX_reg
 
-    -- MMX register stored in ModRM.rm
-    -- (ModRM.mod must equal 3).
+    -- MMX register stored in ModRM.rm (ModRM.mod must equal 3).
   , (,) "N"    $ RG_MMX_rm
    -- Register operand stored in ModRM.rm (ModRM.mod must equal 3)
   , (,) "R"    $ OpType ModRM_rm_mod3 RDQSize
@@ -787,11 +376,9 @@ operandHandlerMap = Map.fromList
   , (,) "MwRd" $ MXRX WSize DSize
   , (,) "MwRy" $ MXRX WSize YSize
 
-    -- Far Pointer (which is an address) stored in ModRM.rm
-    -- (ModRM.mod must not equal 3)
+    -- Far Pointer stored in ModRM.rm (ModRM.mod must not equal 3)
   , (,) "Fv"   $ M_FP
-    -- Memory value stored in ModRM.rm
-    -- (ModRM.mod must not equal 3)
+    -- Memory value stored in ModRM.rm (ModRM.mod must not equal 3)
   , (,) "M"    $ M
   , (,) "Mo"   $ M_X OSize
     -- A reference to an 80-bit floating point value
@@ -801,8 +388,7 @@ operandHandlerMap = Map.fromList
   , (,) "M32fp" $ M_FloatingPoint FPSize32
   , (,) "M64fp" $ M_FloatingPoint FPSize64
 
-
-    -- FP Register indicies
+    -- FP Register indices
   , (,) "ST0"   $ RG_ST 0
   , (,) "ST1"   $ RG_ST 1
   , (,) "ST2"   $ RG_ST 2
@@ -849,8 +435,7 @@ operandHandlerMap = Map.fromList
 
     -- An immediate byte that is sign extended to an operator size.
   , (,) "sIb" $ IM_SB
-    -- An immediate ZSize value that is sign extended to the operator
-    -- size.
+    -- An immediate ZSize value that is sign extended to the operator size.
   , (,) "sIz" $ IM_SZ
 
     -- XMM
@@ -879,94 +464,5 @@ operandHandlerMap = Map.fromList
 lookupOperandType :: MF.MonadFail m => BS.ByteString -> String -> m OperandType
 lookupOperandType i nm =
   case Map.lookup nm operandHandlerMap of
-    Just h -> pure h
+    Just h  -> pure h
     Nothing -> fail $ "Unknown operand for " ++ BSC.unpack i ++ " named " ++ show nm
-
--- | Reverses a list while ensuring the spine of the list is strictly evaluated.
-strictReverse :: [a] -> [a]
-strictReverse = go []
-  where go :: [a] -> [a] -> [a]
-        go r [] = r
-        go r (a:l) = (go $! (a:r)) l
-
--- | Map over a list strictly.
-strictMapList :: Monad m => (a -> m b) -> [a] -> m [b]
-strictMapList = go []
-  where go :: Monad m => [b] -> (a -> m b) -> [a] -> m [b]
-        go r _ [] = pure $! strictReverse r
-        go r f (a:l) = do
-          b <- f a
-          seq b $ go (b:r) f l
-
--- | Parse a definition, returning 'Nothing' if the definition is not
--- supported by flexdis86 (detected as early as possible).
-parse_def ::
-  BS.ByteString ->
-  [BS.ByteString] ->
-  CPURequirement ->
-  Maybe Vendor ->
-  ElemParser (Maybe Def)
-parse_def nm syns creq v = do
-  checkTag "def"
-  let parse_prefix = fromMaybe [] . fmap words <$> opt "pfx" asText
-  prefix <- parse_prefix
-  opc_text <- required_text "opc"
-  oprndNames <- fromMaybe [] . fmap words <$> opt "opr" asText
-  mode <- parse_mode
-  creq' <- parse_CPURequirement creq
-  v' <- parse_vendor v
-  checkEnd
-  -- All XML consumed. Early exit if creq' is definitively unsupported.
-  -- Base is kept because parse_opcode may still override it (e.g. to AMD_3DNOW).
-  if creq' `notElem` (Base : supportedCPUReqs) then return Nothing else do
-    oprnds <- strictMapList (lookupOperandType nm) oprndNames
-    let d0 = Def { _defMnemonic = nm
-                 , _defMnemonicSynonyms = syns
-                 , _defCPUReq = creq'
-                 , _defVendor = v'
-                 , _modeLimit = AnyMode
-                 , _defMode = mode
-                 , _reqAddrSize = Nothing
-                 , _reqOpSize = Nothing
-                 , _defPrefix = prefix
-                 , _requiredPrefix = Nothing
-                 , _defOpcodes = []
-                 , _requiredMod = Nothing
-                 , _requiredReg = Nothing
-                 , _requiredRM  = Nothing
-                 , _x87ModRM = Nothing
-                 , _vexPrefixes = []
-                 , _defOperands = oprnds
-                 }
-    d <- seq d0 $ flip execStateT d0 $ do
-           mapM_ parse_opcode (words opc_text)
-    return $ if defSupported d then Just d else Nothing
-
--- | Consume and discard all remaining elements in the current context.
-skipRemainingElts :: ElemParser ()
-skipRemainingElts = do
-  me <- getNextElt
-  case me of
-    Nothing -> return ()
-    Just _  -> skipRemainingElts
-
-parse_instruction :: ElemParser [Def]
-parse_instruction = do
-  checkTag "instruction"
-  (mnem, syns) <- parse_mnemonics
-  cpuReq <- parse_CPURequirement Base
-  v <- parse_vendor Nothing
-  if cpuReq `notElem` (Base : supportedCPUReqs)
-    then skipRemainingElts >> return []
-    else catMaybes <$> remainingElts (parse_def mnem syns cpuReq v)
-
-parse_x86_optable :: ElemParser [Def]
-parse_x86_optable = do
-  checkTag "x86optable"
-  concat <$> remainingElts parse_instruction
-
-parseOpTable :: BS.ByteString -> Either String [Def]
-parseOpTable bs = do
-  case parseXMLDoc bs of
-    Nothing -> Left "Not an XML document"
-    Just elt -> runElemParser parse_x86_optable elt
